+++ logs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # 改动日志
+## 2026-05-07 20:52 | auto: 05-07 20:52 | src/contexts/ActionContext.tsx src/screens/SettingsScreen.tsx 
+- src/contexts/ActionContext.tsx
+- src/screens/SettingsScreen.tsx
+
 ## 2026-05-07 20:48 | auto: 05-07 20:48 | src/screens/HomeScreen.tsx 
 - src/screens/HomeScreen.tsx